Moved schemas entry point
