+++ packages/search-types/schemas.d.ts
@@ -1,0 +1,1 @@
+export * from './dist/__tests__/schemas';
