+++ packages/search-types/schemas.js
@@ -1,0 +1,12 @@
+'use strict';
+
+function __export(m) {
+  for (var p in m) {
+    if (!exports.hasOwnProperty(p)) {
+      exports[p] = m[p];
+    }
+  }
+}
+
+Object.defineProperty(exports, '__esModule', { value: true });
+__export(require('./dist/__tests__/schemas'));
