+++ packages/search-types/src/__tests__/schemas/facet.schema.ts
@@ -1,4 +1,4 @@
-import { Facet } from './facet.model';
+import { Facet } from '../../facet/facet.model';
 
 export const FacetSchema: Facet = {
   title: expect.any(String),
