+++ packages/search-types/src/__tests__/schemas/index.ts
@@ -1,0 +1,10 @@
+export * from './banner.schema';
+export * from './facet.schema';
+export * from './filter.schema';
+export * from './promoted.schema';
+export * from './next-query.schema';
+export * from './related-tag.schema';
+export * from './redirection.schema';
+export * from './result.schema';
+export * from './suggestion.schema';
+export * from './tagging.schema';
